refactor: reorganize schema imports and update file structure
