+++ components/auth/register-form.tsx
@@ -6,11 +6,11 @@ import { useActionState } from "react";
 import { TriangleAlert } from "lucide-react";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import InputError from "@/components/stocks/input-error";
+import InputError from "@/components/shared/input-error";
 import { registerCredentials } from "@/actions/auth.action";
-import ButtonSubmit from "@/components/stocks/button-submit";
-import InputStrongPassword from "../stocks/input-strong-password";
-import InputShowPassword from "@/components/stocks/input-show-password";
+import ButtonSubmit from "@/components/shared/button-submit";
+import InputStrongPassword from "@/components/shared/input-strong-password";
+import InputShowPassword from "@/components/shared/input-show-password";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 
 export function RegisterForm() {
