+++ components/shared/input-strong-password.tsx
